backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
asyncpg>=0.29.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
dapr>=1.12.0
dapr-ext-fastapi>=1.12.0
cloudevents>=1.10.0
httpx>=0.25.0
openai>=1.0.0
pydantic-settings>=2.0.0
python-jose[cryptography]>=3.3.0
alembic>=1.13.0
email-validator>=2.0.0
fastapi-clerk-auth>=0.0.9
PyJWT[crypto]>=2.8.0
